add skip_restore. NOTE: it will impact view in DAG

+++ src/customizable_continuous_integration/automations/bigquery_archiver/entity/dataset.py
@@ -7,7 +7,7 @@ Revision History:
 ------------------------------------------------------------------------------
   23/02/2025   Ryan, Gao       Initial creation
   03/04/2025   Ryan, Gao       Set project in dataset gcs_prefix
-  10/04/2025   Ryan, Gao       Add archive timestamp to restored dataset labels
+  10/04/2025   Ryan, Gao       Add archive timestamp to dataset labels; Add skip_restore
 """
 
 import datetime
@@ -273,6 +273,9 @@ class BigqueryArchivedDatasetEntity(BigqueryBaseArchiveEntity):
         fully_qualified_identity = self.fully_qualified_identity
         if self.destination_gcp_project_id and self.destination_bigquery_dataset:
             fully_qualified_identity = f"{self.destination_gcp_project_id}.{self.destination_bigquery_dataset}"
+        if restore_config.get("skip_restore", {}).get(self.identity, False):
+            print(f"Skip restoring {self.entity_type} {fully_qualified_identity}")
+            return
         if restore_config.get("overwrite_existing", False):
             bigquery_client.delete_dataset(fully_qualified_identity, delete_contents=True, not_found_ok=True)
         dataset = bigquery_client.create_dataset(fully_qualified_identity, exists_ok=True)
